Moves UserToken to auth.utils, protects start_batch
