+++ chowda/routers/dashboard.py
@@ -1,45 +1,8 @@
-from typing import Annotated
-
-from fastapi import APIRouter, Depends, HTTPException, Request, status
+from fastapi import APIRouter, Depends, Request, status
 from metaflow.integrations import ArgoEvent
-from pydantic import BaseModel, Field
 from starlette.responses import RedirectResponse, Response
 
-from chowda.config import API_AUDIENCE
-
-
-class UserToken(BaseModel):
-    """ID Token model for authorization."""
-
-    name: str
-    email: str | None = None
-    roles: set[str] = Field(set(), alias=f'{API_AUDIENCE}/roles')
-
-
-unauthorized = HTTPException(
-    status_code=status.HTTP_303_SEE_OTHER,
-    detail='Not Authorized',
-    headers={'Location': '/admin'},
-)
-
-
-def user(request: Request):
-    """Get the user token from the session."""
-    user = request.session.get('user', None)
-    if not user:
-        request.session['error'] = 'Not Logged In'
-        raise unauthorized
-    return UserToken(**user)
-
-
-def admin_user(request: Request, user: Annotated[UserToken, Depends(user)]):
-    """Check if the user has the admin role."""
-    if 'admin' not in user.roles:
-        request.session['error'] = 'Not Authorized'
-        raise unauthorized
-
-    return user
-
+from chowda.auth.utils import admin_user
 
 dashboard = APIRouter(dependencies=[Depends(admin_user)])
